code overhaul/submodule update
