--- modules/posix/android
@@ -1,1 +1,0 @@
-android
