+++ modules/posix/concrt.cpp
@@ -6,101 +6,72 @@
 // ---------------------------------------------------------------------------
 #include <coroutine/concrt.h>
 
-#include <atomic>
-#include <chrono>
 #include <cstdio>
 #include <ctime>
-#include <system_error>
 
-#include <pthread.h>
+namespace concrt {
 
 using namespace std;
 using namespace std::chrono;
 
-namespace concrt
-{
-struct latch_posix final
-{
-    atomic<uint32_t> count{};
-    pthread_cond_t cv{};
-    pthread_mutex_t mtx{};
-};
+latch::latch(uint32_t count) noexcept(false) : ref{count}, cv{}, mtx{} {
+    // increase count on ctor phase
 
-auto for_posix(latch* wg) noexcept -> latch_posix*
-{
-    static_assert(sizeof(latch_posix) <= sizeof(latch));
-    return reinterpret_cast<latch_posix*>(wg);
-}
-auto for_posix(const latch* wg) noexcept -> const latch_posix*
-{
-    return for_posix(const_cast<latch*>(wg));
-}
+    // instread of std::mutex, use customized attributes
+    pthread_mutexattr_t attr{};
+    if (auto ec = pthread_mutexattr_init(&attr))
+        throw system_error{ec, system_category(), "pthread_mutexattr_init"};
 
-latch::latch(uint32_t count) noexcept(false) : storage{}
-{
-    auto wg = new (for_posix(this)) latch_posix{};
+    if (auto ec = pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_ERRORCHECK))
+        throw system_error{ec, system_category(), "pthread_mutexattr_settype"};
 
-    // init attr?
-    if (auto ec = pthread_mutex_init(addressof(wg->mtx), nullptr))
+    if (auto ec = pthread_mutexattr_setprotocol(&attr, PTHREAD_PRIO_NONE))
+        throw system_error{ec, system_category(),
+                           "pthread_mutexattr_setprotocol"};
+
+    if (auto ec = pthread_mutex_init(addressof(this->mtx), &attr))
         throw system_error{ec, system_category(), "pthread_mutex_init"};
 
-    if (auto ec = pthread_cond_init(addressof(wg->cv), nullptr))
-        throw system_error{ec, system_category(), "pthread_cond_init"};
+    if (auto ec = pthread_mutexattr_destroy(&attr))
+        throw system_error{ec, system_category(), "pthread_mutexattr_destroy"};
 
-    // just increase count.
-    // notice that delta is always positive
-    wg->count.fetch_add(count);
+    if (auto ec = pthread_cond_init(addressof(this->cv), nullptr))
+        throw system_error{ec, system_category(), "pthread_cond_init"};
 }
 
-latch::~latch() noexcept
-{
-    auto wg = for_posix(this);
-    try
-    {
-        if (auto ec = pthread_cond_destroy(addressof(wg->cv)))
+latch::~latch() noexcept {
+    try {
+        if (auto ec = pthread_cond_destroy(addressof(this->cv)))
             throw system_error{ec, system_category(), "pthread_cond_destroy"};
 
-        if (auto ec = pthread_mutex_destroy(addressof(wg->mtx)))
+        if (auto ec = pthread_mutex_destroy(addressof(this->mtx)))
             throw system_error{ec, system_category(), "pthread_mutex_destroy"};
-    }
-    catch (const system_error& ex)
-    {
+    } catch (const system_error& ex) {
         fputs(ex.what(), stderr);
     }
 }
-void latch::count_down_and_wait() noexcept(false)
-{
+void latch::count_down_and_wait() noexcept(false) {
     this->count_down();
     this->wait();
 }
-
-void latch::count_down(uint32_t n) noexcept(false)
-{
-    auto wg = for_posix(this);
-
+void latch::count_down(uint32_t n) noexcept(false) {
     // underflow
-    if (n > wg->count.load(memory_order_acquire))
+    if (n > this->ref.load(memory_order_acquire))
         throw underflow_error{"latch::count_down"};
 
-    wg->count.fetch_sub(n, memory_order_release);
-    if (wg->count.load(memory_order_acquire) != 0)
+    this->ref.fetch_sub(n, memory_order_release);
+    if (this->ref.load(memory_order_acquire) != 0)
         return;
-
     // notify error
-    if (auto ec = pthread_cond_signal(&wg->cv))
+    if (auto ec = pthread_cond_signal(&this->cv))
         throw system_error{ec, system_category(), "pthread_cond_signal"};
 }
-
-bool latch::is_ready() const noexcept
-{
-    const auto wg = for_posix(this);
-    return wg->count.load(memory_order_acquire) == 0;
+bool latch::is_ready() const noexcept {
+    return this->ref.load(memory_order_acquire) == 0;
 }
-
-std::errc timed_wait(latch_posix& wg, microseconds timeout) noexcept
-{
+std::errc latch::timed_wait(microseconds timeout) noexcept {
     // check count fast
-    if (wg.count.load(memory_order_relaxed) == 0)
+    if (this->ref.load(memory_order_relaxed) == 0)
         return std::errc{};
 
     // `pthread_cond_timedwait` uses absolute time(== time point).
@@ -114,7 +85,7 @@ std::errc timed_wait(latch_posix& wg, microseconds timeout) noexcept
     until.tv_nsec = duration_cast<nanoseconds>(end_time).count() % 1'000'000;
 
 CheckCurrentCount:
-    if (wg.count.load(memory_order_acquire) == 0)
+    if (this->ref.load(memory_order_acquire) == 0)
         return std::errc{};
 
     // This check is for possible spurious wakeup.
@@ -125,11 +96,11 @@ CheckCurrentCount:
 
     int reason = 0;
     {
-        if (auto ec = pthread_mutex_lock(addressof(wg.mtx)))
+        if (auto ec = pthread_mutex_lock(addressof(this->mtx)))
             return std::errc{ec};
-        reason = pthread_cond_timedwait(addressof(wg.cv), addressof(wg.mtx),
-                                        addressof(until));
-        if (auto ec = pthread_mutex_unlock(addressof(wg.mtx)))
+        reason = pthread_cond_timedwait(addressof(this->cv),
+                                        addressof(this->mtx), addressof(until));
+        if (auto ec = pthread_mutex_unlock(addressof(this->mtx)))
             return std::errc{ec};
     }
     if (reason == ETIMEDOUT) // this might be a spurious wakeup
@@ -138,17 +109,12 @@ CheckCurrentCount:
     // reason containes error code at this moment
     return std::errc{reason};
 }
-
-void latch::wait() noexcept(false)
-{
-    auto wg = for_posix(this);
+void latch::wait() noexcept(false) {
     auto timeout = 100ms;
-
-    while (true)
-    {
+    while (true) {
         // since latch doesn't provide interface for waiting with timeout,
         // this code will wait more if it reached timeout
-        auto ec = timed_wait(*wg, timeout);
+        auto ec = this->timed_wait(timeout);
         if (ec == std::errc{})
             return;
         if (ec != errc::timed_out)
@@ -157,17 +123,4 @@ void latch::wait() noexcept(false)
         // ... update timeout if it need to be changed ...
     }
 }
-
-// barrier::barrier(uint32_t num_threads) noexcept(false) : storage{}
-// {
-//     throw runtime_error{"not implemented"};
-// }
-// barrier::~barrier() noexcept
-// {
-// }
-// void barrier::arrive_and_wait() noexcept
-// {
-//     throw runtime_error{"not implemented"};
-// }
-
 } // namespace concrt
