+++ modules/windows.vcxproj
@@ -23,7 +23,6 @@
     <ClCompile Include="windows\concrt.cpp" />
     <ClCompile Include="windows\dllmain.cpp" />
     <ClCompile Include="windows\net.cpp" />
-    <ClCompile Include="windows\suspend.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\interface\coroutine\channel.hpp" />
