+++ modules/windows.vcxproj.filters
@@ -25,9 +25,6 @@
     <ClCompile Include="windows\net.cpp">
       <Filter>windows</Filter>
     </ClCompile>
-    <ClCompile Include="windows\suspend.cpp">
-      <Filter>windows</Filter>
-    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\interface\coroutine\channel.hpp">
