+++ modules/windows/concrt.cpp
@@ -17,65 +17,38 @@
 
 #include "stop_watch.hpp"
 
-namespace concrt
-{
+namespace concrt {
 using namespace std;
 using namespace std::experimental;
 using namespace std::chrono;
 
-class alertable_event
-{
-    HANDLE ev{};
-
-  public:
-    alertable_event() noexcept(false);
-    ~alertable_event() noexcept;
-    alertable_event(const alertable_event&) = delete;
-    alertable_event(alertable_event&&) = delete;
-    alertable_event& operator=(const alertable_event&) = delete;
-    alertable_event& operator=(alertable_event&&) = delete;
-
-    bool is_closed() const noexcept;
-    void close() noexcept;
-    void reset() noexcept;
-    void set() noexcept;
-    bool wait(uint32_t ms) noexcept;
-};
-
-alertable_event::alertable_event() noexcept(false)
-    : ev{CreateEventA(nullptr, false, false, nullptr)}
-{
-    if (ev == INVALID_HANDLE_VALUE)
+win32_event::win32_event() noexcept(false)
+    : native{CreateEventA(nullptr, false, false, nullptr)} {
+    if (native == INVALID_HANDLE_VALUE)
         throw system_error{gsl::narrow_cast<int>(GetLastError()),
                            system_category(), "CreateEventA"};
 }
-alertable_event::~alertable_event() noexcept
-{
+win32_event::~win32_event() noexcept {
     this->close();
 }
-bool alertable_event::is_closed() const noexcept
-{
-    return ev == INVALID_HANDLE_VALUE;
+bool win32_event::is_closed() const noexcept {
+    return native == INVALID_HANDLE_VALUE;
 }
-void alertable_event::close() noexcept
-{
-    CloseHandle(ev);
-    ev = INVALID_HANDLE_VALUE;
+void win32_event::close() noexcept {
+    CloseHandle(native);
+    native = INVALID_HANDLE_VALUE;
 }
-void alertable_event::reset() noexcept
-{
-    ResetEvent(ev);
+void win32_event::reset() noexcept {
+    ResetEvent(native);
 }
-void alertable_event::set() noexcept
-{
-    SetEvent(ev);
+void win32_event::set() noexcept {
+    SetEvent(native);
 }
-bool alertable_event::wait(uint32_t ms) noexcept
-{
+bool win32_event::wait(uint32_t ms) noexcept {
     static_assert(WAIT_OBJECT_0 == 0);
 
     // This makes APC available. expecially for Overlapped I/O
-    if (WaitForSingleObjectEx(ev, ms, TRUE))
+    if (WaitForSingleObjectEx(native, ms, TRUE))
         // WAIT_FAILED	: use GetLastError in the case
         // WAIT_TIMEOUT	: this is expected. user can try again
         // WAIT_IO_COMPLETION : return because of APC
@@ -87,78 +60,51 @@ bool alertable_event::wait(uint32_t ms) noexcept
     return true;
 }
 
-// this cast doesn't harm semantics
-#define DECLARE_LATCH_MEMBERS()                                                \
-    auto ev = (alertable_event*)(storage + 0);                                 \
-    auto ctr = (atomic_uint64_t*)(storage + sizeof(*ev));
-
-latch::latch(uint32_t delta) noexcept(false) : storage{}
-{
-    DECLARE_LATCH_MEMBERS();
-
-    if (new (ctr) atomic_uint64_t{})
-        ctr->fetch_add(delta);
-
-    if (new (ev) alertable_event{})
-        ev->reset();
+latch::latch(uint32_t delta) noexcept(false) : ev{}, ref{delta} {
+    ev.reset();
 }
-GSL_SUPPRESS(con .4)
-latch::~latch() noexcept
-{
-    DECLARE_LATCH_MEMBERS();
-    UNREFERENCED_PARAMETER(ctr);
-    ev->~alertable_event();
-}
-void latch::count_down_and_wait() noexcept(false)
-{
+
+void latch::count_down_and_wait() noexcept(false) {
     this->count_down();
     this->wait();
 }
-void latch::count_down(uint32_t n) noexcept(false)
-{
-    DECLARE_LATCH_MEMBERS();
+void latch::count_down(uint32_t n) noexcept(false) {
 
-    if (ctr->load(memory_order_acquire) < n)
+    if (ref.load(memory_order_acquire) < n)
         throw underflow_error{"latch's count can't be negative"};
 
-    ctr->fetch_sub(n, memory_order_release);
-    if (ctr->load(memory_order_acquire) == 0)
-        ev->set();
+    ref.fetch_sub(n, memory_order_release);
+    if (ref.load(memory_order_acquire) == 0)
+        ev.set();
 }
 GSL_SUPPRESS(f .23)
 GSL_SUPPRESS(con .4)
-bool latch::is_ready() const noexcept
-{
-    DECLARE_LATCH_MEMBERS();
-    if (ctr->load() == 0)
+bool latch::is_ready() const noexcept {
+    if (ref.load(memory_order_acquire) == 0)
         // if it is not closed, test it
-        return ev->is_closed() ? true : ev->wait(0);
+        return ev.is_closed() ? true : ev.wait(0);
     return false;
 }
-void latch::wait() noexcept(false)
-{
-    DECLARE_LATCH_MEMBERS();
-    UNREFERENCED_PARAMETER(ctr);
-    while (ev->wait(INFINITE) == false)
-        ; // standard interface doesn't define timed wait.
+void latch::wait() noexcept(false) {
+    // standard interface doesn't define timed wait.
+    while (ev.wait(INFINITE) == false)
+        ;
 }
 
 void ptp_work::resume_on_thread_pool(PTP_CALLBACK_INSTANCE, //
-                                     PVOID ctx, PTP_WORK work)
-{
+                                     PVOID ctx, PTP_WORK work) {
     if (auto coro = coroutine_handle<void>::from_address(ctx))
         if (coro.done() == false)
             coro.resume();
 
     ::CloseThreadpoolWork(work); // one-time work item
 }
-auto ptp_work::suspend(coroutine_handle<void> coro) noexcept -> uint32_t
-{
+auto ptp_work::suspend(coroutine_handle<void> coro) noexcept -> uint32_t {
     // just make sure no data loss in `static_cast`
     static_assert(sizeof(uint32_t) == sizeof(DWORD));
 
-    auto work = ::CreateThreadpoolWork(resume_on_thread_pool, coro.address(),
-                                       nullptr);
+    auto work =
+        ::CreateThreadpoolWork(resume_on_thread_pool, coro.address(), nullptr);
     if (work == nullptr)
         return GetLastError();
 
@@ -166,24 +112,19 @@ auto ptp_work::suspend(coroutine_handle<void> coro) noexcept -> uint32_t
     return S_OK;
 }
 
-section::section() noexcept(false)
-{
+section::section() noexcept(false) {
     InitializeCriticalSectionAndSpinCount(this, 0600);
 }
-section::~section() noexcept
-{
+section::~section() noexcept {
     DeleteCriticalSection(this);
 }
-bool section::try_lock() noexcept
-{
+bool section::try_lock() noexcept {
     return TryEnterCriticalSection(this);
 }
-void section::lock() noexcept(false)
-{
+void section::lock() noexcept(false) {
     EnterCriticalSection(this);
 }
-void section::unlock() noexcept(false)
-{
+void section::unlock() noexcept(false) {
     LeaveCriticalSection(this);
 }
 } // namespace concrt
