+++ modules/windows/net.cpp
@@ -9,8 +9,7 @@
 using namespace std;
 
 auto wait_io_tasks(chrono::nanoseconds) noexcept(false)
-    -> coro::enumerable<io_task_t>
-{
+    -> coro::enumerable<io_task_t> {
     // windows implementation rely on callback.
     // So this function will yield nothing
     co_return;
@@ -19,8 +18,7 @@ auto wait_io_tasks(chrono::nanoseconds) noexcept(false)
 GSL_SUPPRESS(type .1)
 GSL_SUPPRESS(f .6)
 void CALLBACK onWorkDone(DWORD errc, DWORD sz, LPWSAOVERLAPPED pover,
-                         DWORD flags) noexcept
-{
+                         DWORD flags) noexcept {
     UNREFERENCED_PARAMETER(flags);
     io_work_t* work = reinterpret_cast<io_work_t*>(pover);
 
@@ -32,28 +30,24 @@ void CALLBACK onWorkDone(DWORD errc, DWORD sz, LPWSAOVERLAPPED pover,
     work->task.resume();
 }
 
-bool io_work_t::ready() const noexcept
-{
+bool io_work_t::ready() const noexcept {
     return false; // trigger `await_suspend`
 }
 
-uint32_t io_work_t::error() const noexcept
-{
+uint32_t io_work_t::error() const noexcept {
     static_assert(sizeof(DWORD) == sizeof(uint32_t));
     return gsl::narrow_cast<uint32_t>(Internal);
 }
 
 auto zero_overlapped(gsl::not_null<io_work_t*> work) noexcept
-    -> gsl::not_null<OVERLAPPED*>
-{
+    -> gsl::not_null<OVERLAPPED*> {
     OVERLAPPED* pover = work;
     *pover = OVERLAPPED{};
     return pover;
 }
 
 GSL_SUPPRESS(type .1)
-auto make_wsa_buf(buffer_view_t v) noexcept -> WSABUF
-{
+auto make_wsa_buf(buffer_view_t v) noexcept -> WSABUF {
     WSABUF buf{}; // expect NRVO
     buf.buf = reinterpret_cast<char*>(v.data());
     buf.len = gsl::narrow_cast<ULONG>(v.size_bytes());
@@ -63,8 +57,7 @@ auto make_wsa_buf(buffer_view_t v) noexcept -> WSABUF
 GSL_SUPPRESS(type .1)
 GSL_SUPPRESS(type .3)
 auto send_to(uint64_t sd, const sockaddr_in6& remote, buffer_view_t buffer,
-             io_work_t& work) noexcept(false) -> io_send_to&
-{
+             io_work_t& work) noexcept(false) -> io_send_to& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -80,8 +73,7 @@ auto send_to(uint64_t sd, const sockaddr_in6& remote, buffer_view_t buffer,
 GSL_SUPPRESS(type .1)
 GSL_SUPPRESS(type .3)
 auto send_to(uint64_t sd, const sockaddr_in& remote, buffer_view_t buffer,
-             io_work_t& work) noexcept(false) -> io_send_to&
-{
+             io_work_t& work) noexcept(false) -> io_send_to& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -94,8 +86,7 @@ auto send_to(uint64_t sd, const sockaddr_in& remote, buffer_view_t buffer,
     return *reinterpret_cast<io_send_to*>(addressof(work));
 }
 
-void io_send_to::suspend(io_task_t rh) noexcept(false)
-{
+void io_send_to::suspend(io_task_t rh) noexcept(false) {
     const auto sd = gsl::narrow_cast<SOCKET>(Internal);
     const auto addr = addressof(this->ep->addr);
     const auto addrlen = gsl::narrow_cast<socklen_t>(InternalHigh);
@@ -114,15 +105,13 @@ void io_send_to::suspend(io_task_t rh) noexcept(false)
     throw system_error{ec, system_category(), "WSASendTo"};
 }
 
-int64_t io_send_to::resume() noexcept
-{
+int64_t io_send_to::resume() noexcept {
     return gsl::narrow_cast<int64_t>(InternalHigh);
 }
 
 GSL_SUPPRESS(type .1)
 auto recv_from(uint64_t sd, sockaddr_in6& remote, buffer_view_t buffer,
-               io_work_t& work) noexcept(false) -> io_recv_from&
-{
+               io_work_t& work) noexcept(false) -> io_recv_from& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -137,8 +126,7 @@ auto recv_from(uint64_t sd, sockaddr_in6& remote, buffer_view_t buffer,
 
 GSL_SUPPRESS(type .1)
 auto recv_from(uint64_t sd, sockaddr_in& remote, buffer_view_t buffer,
-               io_work_t& work) noexcept(false) -> io_recv_from&
-{
+               io_work_t& work) noexcept(false) -> io_recv_from& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -151,8 +139,7 @@ auto recv_from(uint64_t sd, sockaddr_in& remote, buffer_view_t buffer,
     return *reinterpret_cast<io_recv_from*>(addressof(work));
 }
 
-void io_recv_from::suspend(io_task_t rh) noexcept(false)
-{
+void io_recv_from::suspend(io_task_t rh) noexcept(false) {
     const auto sd = gsl::narrow_cast<SOCKET>(Internal);
     auto addr = addressof(this->ep->addr);
     auto addrlen = gsl::narrow_cast<socklen_t>(InternalHigh);
@@ -171,15 +158,13 @@ void io_recv_from::suspend(io_task_t rh) noexcept(false)
     throw system_error{ec, system_category(), "WSARecvFrom"};
 }
 
-int64_t io_recv_from::resume() noexcept
-{
+int64_t io_recv_from::resume() noexcept {
     return gsl::narrow_cast<int64_t>(InternalHigh);
 }
 
 GSL_SUPPRESS(type .1)
 auto send_stream(uint64_t sd, buffer_view_t buffer, uint32_t flag,
-                 io_work_t& work) noexcept(false) -> io_send&
-{
+                 io_work_t& work) noexcept(false) -> io_send& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -190,8 +175,7 @@ auto send_stream(uint64_t sd, buffer_view_t buffer, uint32_t flag,
     return *reinterpret_cast<io_send*>(addressof(work));
 }
 
-void io_send::suspend(io_task_t rh) noexcept(false)
-{
+void io_send::suspend(io_task_t rh) noexcept(false) {
     const auto sd = gsl::narrow_cast<SOCKET>(Internal);
     const auto flag = gsl::narrow_cast<DWORD>(InternalHigh);
     auto buf = make_wsa_buf(buffer);
@@ -207,15 +191,13 @@ void io_send::suspend(io_task_t rh) noexcept(false)
     throw system_error{ec, system_category(), "WSASend"};
 }
 
-int64_t io_send::resume() noexcept
-{
+int64_t io_send::resume() noexcept {
     return gsl::narrow_cast<int64_t>(InternalHigh);
 }
 
 GSL_SUPPRESS(type .1)
 auto recv_stream(uint64_t sd, buffer_view_t buffer, uint32_t flag,
-                 io_work_t& work) noexcept(false) -> io_recv&
-{
+                 io_work_t& work) noexcept(false) -> io_recv& {
     static_assert(sizeof(SOCKET) == sizeof(uint64_t));
     static_assert(sizeof(HANDLE) == sizeof(SOCKET));
 
@@ -226,8 +208,7 @@ auto recv_stream(uint64_t sd, buffer_view_t buffer, uint32_t flag,
     return *reinterpret_cast<io_recv*>(addressof(work));
 }
 
-void io_recv::suspend(io_task_t rh) noexcept(false)
-{
+void io_recv::suspend(io_task_t rh) noexcept(false) {
     const auto sd = gsl::narrow_cast<SOCKET>(Internal);
     auto flag = gsl::narrow_cast<DWORD>(InternalHigh);
     auto buf = make_wsa_buf(buffer);
@@ -243,7 +224,6 @@ void io_recv::suspend(io_task_t rh) noexcept(false)
     throw system_error{ec, system_category(), "WSARecv"};
 }
 
-int64_t io_recv::resume() noexcept
-{
+int64_t io_recv::resume() noexcept {
     return gsl::narrow_cast<int64_t>(InternalHigh);
 }
