+++ test/vs_resolver.cpp
@@ -1,0 +1,167 @@
+//
+//  Author  : github.com/luncliff (luncliff@gmail.com)
+//  License : CC BY 4.0
+//
+#include "./socket_test.h"
+
+#include <CppUnitTest.h>
+#include <sdkddkver.h>
+
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_MODULE_INITIALIZE(winsock_init)
+{
+    load_network_api();
+}
+
+class resolver_name_info_test : public TestClass<resolver_name_info_test>
+{
+    TEST_METHOD(get_name_current_host)
+    {
+        const auto name = host_name();
+        Assert::IsNotNull(name); // non-null
+    }
+
+    TEST_METHOD(get_name_ipv6_address)
+    {
+        auto buffer = std::make_unique<char[]>(NI_MAXHOST);
+        auto name = buffer.get();
+
+        endpoint_t ep{};
+        ep.in6.sin6_family = AF_INET6;
+        ep.in6.sin6_addr = in6addr_any;
+        ep.in6.sin6_port = htons(7654);
+
+        // non-zero for error. it is redirected from getnameinfo
+        auto ec = get_name(ep, name);
+        Assert::IsTrue(ec == NO_ERROR);
+    }
+
+    TEST_METHOD(get_name_ipv4_address)
+    {
+        auto buffer = std::make_unique<char[]>(NI_MAXHOST);
+        auto name = buffer.get();
+
+        endpoint_t ep{};
+        ep.in4.sin_family = AF_INET;
+        ep.in4.sin_addr.s_addr = INADDR_ANY;
+        ep.in4.sin_port = htons(7654);
+
+        // non-zero for error. it is redirected from getnameinfo
+        auto ec = get_name(ep, name);
+        Assert::IsTrue(ec == NO_ERROR);
+    }
+};
+
+class resolver_addr_info_test : public TestClass<resolver_addr_info_test>
+{
+    addrinfo hint{};
+    size_t count = 0u;
+
+    TEST_METHOD(get_addr_tcpv6_for_connect)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_STREAM;
+        hint.ai_flags = AI_ALL | AI_V4MAPPED | AI_NUMERICHOST;
+
+        for (auto ep : resolve(hint, "::1", "echo"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(7));
+            auto loopback = IN6_IS_ADDR_LOOPBACK(&ep.in6.sin6_addr);
+            Assert::IsTrue(loopback);
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+
+    TEST_METHOD(get_addr_tcpv6_for_listen_text)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_STREAM;
+
+        hint.ai_flags = AI_PASSIVE | AI_V4MAPPED;
+        for (auto ep : resolve(hint, nullptr, "https"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(443));
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+
+    TEST_METHOD(get_addr_tcpv6_for_listen_numeric)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_STREAM;
+        hint.ai_flags = AI_PASSIVE | AI_V4MAPPED //
+                        | AI_NUMERICSERV | AI_NUMERICHOST;
+
+        for (auto ep : resolve(hint, "fe80::", "57132"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(57132));
+            // error message from macro is too long. make it short
+            bool linklocal = IN6_IS_ADDR_LINKLOCAL(&ep.in6.sin6_addr);
+            Assert::IsTrue(linklocal);
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+
+    TEST_METHOD(get_addr_udpv6_for_bind)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_DGRAM;
+        hint.ai_flags = AI_ALL | AI_V4MAPPED //
+                        | AI_NUMERICHOST | AI_NUMERICSERV;
+
+        for (auto ep : resolve(hint, "::", "9283"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(9283));
+            auto unspec = IN6_IS_ADDR_UNSPECIFIED(&ep.in6.sin6_addr);
+            Assert::IsTrue(unspec);
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+
+    TEST_METHOD(get_addr_ipv6_for_bind)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_RAW;
+        hint.ai_flags = AI_ALL | AI_V4MAPPED //
+                        | AI_NUMERICHOST | AI_NUMERICSERV;
+
+        for (auto ep : resolve(hint, "::0.0.0.0", "9287"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(9287));
+
+            auto unspec = IN6_IS_ADDR_UNSPECIFIED(&ep.in6.sin6_addr);
+            Assert::IsTrue(unspec);
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+
+    TEST_METHOD(get_addr_ipv6_for_multicast)
+    {
+        hint.ai_family = AF_INET6;
+        hint.ai_socktype = SOCK_RAW;
+        hint.ai_flags = AI_ALL | AI_NUMERICHOST | AI_NUMERICSERV;
+
+        for (auto ep : resolve(hint, "FF0E::1", "7283"))
+        {
+            Assert::IsTrue(ep.in6.sin6_family == AF_INET6);
+            Assert::IsTrue(ep.in6.sin6_port == htons(7283));
+
+            auto global = IN6_IS_ADDR_MC_GLOBAL(&ep.in6.sin6_addr);
+            Assert::IsTrue(global);
+
+            ++count;
+        }
+        Assert::IsTrue(count > 0);
+    }
+};
